refactor(subagent): extract execute() into subagent/ modules
The Agent-tool executor's execute() was ~673 LOC — over half the file —
linearly threading seven concerns through one try/finally. Extract each
concern into a focused module under src/agent/tools/subagent/, leaving
execute() a thin orchestrator.

- input-parse.ts           AgentInput shape + parseAgentInput validator (pure)
- child-config.ts          child AgentSession config + nesting/depth wiring
- background-branch.ts     mode:'background' register-and-return path
- foreground-promotion.ts  foreground run + Ctrl+B promotion race + cleanup
- failure-payload.ts       telemetry wrapper + structured failure-payload shaping

Behavior-preserving: pure move + wire, public API unchanged.
subagent-executor.ts 1316 -> 577 LOC. tsc --noEmit clean; all 86
subagent-executor.test.ts cases pass.

Module params sourced from optional fields (IAgentSession.sessionId,
AgentInput.id_prefix) are typed string | undefined to preserve the
pre-extraction contract — they flow only into telemetry/registry, which
already accept undefined.
